+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -20,6 +20,8 @@ CHANGELOG
    use the `router.request_context.base_url` parameter or the `framework.router.default_uri` config option instead
  * Add decoration stack to `debug:container` command output
  * Add `MicroKernelTrait::getAllowedEnvs()` to list and enforce allowed values for `APP_ENV`
+ * Change `caching.max_ttl` option to default to `86400` (24h) instead of `null`
+ * Deprecate setting `caching.max_ttl` to `null`, use a positive integer instead
 
 8.0
 ---